+++ Readme/Changelog.md
@@ -7,6 +7,7 @@
 3. More and updated temporary rules
 4. Updated documentation, specifically main readme.md, FirewallParameters.md and more PS command samples.
 5. Add functions to query SQL Instances on network and SQL related software cmdlets.
+6. Move configuration specific global variables into a separate file
 
 # Bugfixes
 1. Executing rules for store apps would fail for new Windows accounts
@@ -18,3 +19,4 @@
 1. Make all files use tabs instead of spaces
 2. Expand modules to include correct comments, help and manifest files
 3. A few more TODO's, notes and relevant comments
+4. Reorganize and split scripts with rules for multiple targets